Removed extra line before main fn (#272)
* Removed extra line before main fn

* removed extra line in trainers.py

* - Refactored store.js into two files: store.js and codegen.js
- fixed code rendering issue with additional new lines
