+++ CONTRIBUTING.md
@@ -27,6 +27,8 @@ To contribute to Code-Generator App, you will need Nodejs LTS v16, VSCode, Vetur
 
 - Create a virtual environment for python.
 
+  - For docker users please see [this guide](docker/readme.md)
+
   - With pip:
 
     ```sh
@@ -104,9 +106,9 @@ To add a new template,
 
   ```sh
   # format
-  pnpm run fmt
+  pnpm fmt
   # lint
-  pnpm run lint
+  pnpm source_lint
   ```
 
 - To ensure the codebase complies with a style guide, we use black and ufmt to format and check codebase for compliance with PEP8. Install and run with:
@@ -117,7 +119,7 @@ To add a new template,
   # format the codes
   bash scripts/run_code_style.sh fmt
   # lint the codes
-  bash scripts/run_code_style.sh lint
+  bash scripts/run_code_style.sh source_lint
   ```
 
 _NOTE: Even if you have a half-completed/working PR, sending a PR is still a valid contribution and we can help you finish the PR._
