+++ package.json
@@ -9,8 +9,8 @@
     "test:ci": "start-server-and-test --expect 200 serve http://127.0.0.1:5000 test",
     "release": "node scripts/release.js",
     "fmt": "prettier --write . && bash scripts/run_code_style.sh fmt",
-    "min_lint": "prettier --check . && bash scripts/run_code_style.sh min_lint",
-    "lint": "prettier --check . && bash scripts/run_code_style.sh lint"
+    "source_lint": "prettier --check . && bash scripts/run_code_style.sh source_lint",
+    "dist_lint": "prettier --check . && bash scripts/run_code_style.sh dist_lint"
   },
   "dependencies": {
     "@iconify/iconify": "^3.1.0",
