+++ scripts/run_code_style.sh
@@ -2,12 +2,16 @@
 
 set -xeu
 
-if [ $1 == "lint" ]; then
+if [ $1 == "dist_lint" ]; then
     # Check that ./dist-tests/ exists and code is unzipped
     ls ./dist-tests/vision-classification-all/main.py
+    # Comment dist-tests in .gitignore to make black running on ./dist-tests folder
+    # TODO:
     ufmt diff .
     flake8 --select F401,F821 ./dist-tests  # find unused imports and non imported objects
-elif [ $1 == "min_lint" ]; then
+    # Restore .gitignore
+    # TODO:
+elif [ $1 == "source_lint" ]; then
     ufmt diff .
 elif [ $1 == "fmt" ]; then
     ufmt format .
