+++ src/templates/template-text-classification/main.py
@@ -1,23 +1,18 @@
 import os
 from pprint import pformat
 from shutil import copy
-from typing import Any, cast
+from typing import Any
 
 import ignite.distributed as idist
 from data import setup_data
 from ignite.engine import Events
-from ignite.handlers import LRScheduler, PiecewiseLinear
+from ignite.handlers import PiecewiseLinear
 from ignite.metrics import Accuracy, Loss
 from ignite.utils import manual_seed
 from models import TransformerModel
 from torch import nn, optim
 from trainers import setup_evaluator, setup_trainer
 from utils import *
-
-try:
-    from torch.optim.lr_scheduler import _LRScheduler as PyTorchLRScheduler
-except ImportError:
-    from torch.optim.lr_scheduler import LRScheduler as PyTorchLRScheduler
 
 os.environ["TOKENIZERS_PARALLELISM"] = "false"  # remove tokenizer paralleism warning
 
@@ -83,18 +78,11 @@ def run(local_rank: int, config: Any):
     logger.info("Configuration: \n%s", pformat(vars(config)))
     trainer.logger = evaluator.logger = logger
 
-    if isinstance(lr_scheduler, PyTorchLRScheduler):
-        trainer.add_event_handler(
-            Events.ITERATION_COMPLETED,
-            lambda engine: cast(PyTorchLRScheduler, lr_scheduler).step(),
-        )
-    elif isinstance(lr_scheduler, LRScheduler):
-        trainer.add_event_handler(Events.ITERATION_COMPLETED, lr_scheduler)
-    else:
-        trainer.add_event_handler(Events.ITERATION_STARTED, lr_scheduler)
+    trainer.add_event_handler(Events.ITERATION_COMPLETED, lr_scheduler)
+
+    #::: if (it.save_training || it.save_evaluation) { :::#
 
     # setup ignite handlers
-    #::: if (it.save_training || it.save_evaluation) { :::#
     #::: if (it.save_training) { :::#
     to_save_train = {
         "model": model,
@@ -118,6 +106,7 @@ def run(local_rank: int, config: Any):
     #::: } :::#
 
     #::: if (it.logger) { :::#
+
     # experiment tracking
     if rank == 0:
         exp_logger = setup_exp_logging(config, trainer, optimizer, evaluator)
@@ -155,12 +144,14 @@ def run(local_rank: int, config: Any):
     )
 
     #::: if (it.logger) { :::#
+
     # close logger
     if rank == 0:
         exp_logger.close()
     #::: } :::#
-    #
+
     #::: if (it.save_training || it.save_evaluation) { :::#
+
     # show last checkpoint names
     logger.info(
         "Last training checkpoint name - %s",
