+++ src/templates/template-text-classification/trainers.py
@@ -47,10 +47,11 @@ def setup_trainer(
         engine.state.metrics = metric
         return metric
 
-    #
     #::: if(it.deterministic) { :::#
+
     trainer = DeterministicEngine(train_function)
     #::: } else { :::#
+
     trainer = Engine(train_function)
     #::: } :::#
