+++ src/templates/template-vision-classification/trainers.py
@@ -38,10 +38,11 @@ def setup_trainer(
         }
         return {"train_loss": train_loss}
 
-    #
     #::: if(it.deterministic) { :::#
+
     trainer = DeterministicEngine(train_function)
     #::: } else { :::#
+
     trainer = Engine(train_function)
     #::: } :::#
