+++ src/templates/template-vision-dcgan/trainers.py
@@ -87,10 +87,11 @@ def setup_trainer(
 
         return metrics
 
-    #
     #::: if(it.deterministic) { :::#
+
     trainer = DeterministicEngine(train_function)
     #::: } else { :::#
+
     trainer = Engine(train_function)
     #::: } :::#
