+++ src/templates/template-vision-segmentation/main.py
@@ -87,8 +87,9 @@ def run(local_rank: int, config: Any):
     else:
         trainer.add_event_handler(Events.ITERATION_STARTED, lr_scheduler)
 
-    # setup ignite handlers
     #::: if (it.save_training || it.save_evaluation) { :::#
+
+    # setup ignite handlers
     #::: if (it.save_training) { :::#
     to_save_train = {
         "model": model,
@@ -112,6 +113,7 @@ def run(local_rank: int, config: Any):
     #::: } :::#
 
     #::: if (it.logger) { :::#
+
     # experiment tracking
     if rank == 0:
         exp_logger = setup_exp_logging(config, trainer, optimizer, evaluator)
@@ -177,12 +179,14 @@ def run(local_rank: int, config: Any):
     )
 
     #::: if (it.logger) { :::#
+
     # close logger
     if rank == 0:
         exp_logger.close()
     #::: } :::#
-    #
+
     #::: if (it.save_training || it.save_evaluation) { :::#
+
     # show last checkpoint names
     logger.info(
         "Last training checkpoint name - %s",
