+++ src/templates/template-vision-segmentation/trainers.py
@@ -42,10 +42,11 @@ def setup_trainer(
         engine.state.metrics = metric
         return metric
 
-    #
     #::: if(it.deterministic) { :::#
+
     trainer = DeterministicEngine(train_function)
     #::: } else { :::#
+
     trainer = Engine(train_function)
     #::: } :::#
